Refactor HttpUtils to message_layer.Json
