+++ src/main/java/com/streamr/client/protocol/message_layer/BigDecimalAdapter.java
@@ -1,0 +1,19 @@
+package com.streamr.client.protocol.message_layer;
+
+import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.JsonReader;
+import com.squareup.moshi.JsonWriter;
+import java.io.IOException;
+import java.math.BigDecimal;
+
+public class BigDecimalAdapter extends JsonAdapter<BigDecimal> {
+  @Override
+  public BigDecimal fromJson(JsonReader reader) throws IOException {
+    return new BigDecimal(reader.nextString());
+  }
+
+  @Override
+  public void toJson(JsonWriter writer, BigDecimal value) throws IOException {
+    writer.value(value.doubleValue());
+  }
+}
