+++ src/main/java/com/streamr/client/protocol/message_layer/GroupKeyAnnounceAdapter.java
@@ -6,14 +6,13 @@ import com.squareup.moshi.JsonWriter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
 import com.streamr.client.utils.EncryptedGroupKey;
-import com.streamr.client.utils.HttpUtils;
 import java.io.IOException;
 import java.util.List;
 import javax.annotation.Nullable;
 
 public class GroupKeyAnnounceAdapter extends AbstractGroupKeyMessageAdapter<GroupKeyAnnounce> {
   private static final Moshi MOSHI =
-      HttpUtils.addDefaultAdapters
+      Json.addDefaultAdapters
           .apply(new Moshi.Builder())
           .add(EncryptedGroupKey.class, new EncryptedGroupKeyAdapter())
           .build();
