+++ src/main/java/com/streamr/client/protocol/message_layer/GroupKeyErrorResponseAdapter.java
@@ -2,7 +2,6 @@ package com.streamr.client.protocol.message_layer;
 
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.utils.HttpUtils;
 import java.io.IOException;
 import java.util.List;
 import javax.annotation.Nullable;
@@ -22,7 +21,7 @@ public class GroupKeyErrorResponseAdapter
     String streamId = reader.nextString();
     String errorCode = reader.nextString();
     String errorMessage = reader.nextString();
-    List<String> groupKeyIds = HttpUtils.listAdapter.fromJson(reader);
+    List<String> groupKeyIds = Json.listAdapter.fromJson(reader);
     reader.endArray();
 
     return new GroupKeyErrorResponse(requestId, streamId, errorCode, errorMessage, groupKeyIds);
@@ -36,7 +35,7 @@ public class GroupKeyErrorResponseAdapter
     writer.value(message.getStreamId());
     writer.value(message.getCode());
     writer.value(message.getMessage());
-    HttpUtils.listAdapter.toJson(writer, message.getGroupKeyIds());
+    Json.listAdapter.toJson(writer, message.getGroupKeyIds());
     writer.endArray();
   }
 }
