+++ src/main/java/com/streamr/client/protocol/message_layer/GroupKeyRequestAdapter.java
@@ -2,7 +2,6 @@ package com.streamr.client.protocol.message_layer;
 
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.utils.HttpUtils;
 import java.io.IOException;
 import java.util.List;
 import javax.annotation.Nullable;
@@ -20,7 +19,7 @@ public class GroupKeyRequestAdapter extends AbstractGroupKeyMessageAdapter<Group
     String requestId = reader.nextString();
     String streamId = reader.nextString();
     String rsaPublicKey = reader.nextString();
-    List<String> groupKeyIds = HttpUtils.listAdapter.fromJson(reader);
+    List<String> groupKeyIds = Json.listAdapter.fromJson(reader);
     reader.endArray();
 
     return new GroupKeyRequest(requestId, streamId, rsaPublicKey, groupKeyIds);
@@ -32,7 +31,7 @@ public class GroupKeyRequestAdapter extends AbstractGroupKeyMessageAdapter<Group
     writer.value(message.getRequestId());
     writer.value(message.getStreamId());
     writer.value(message.getPublicKey());
-    HttpUtils.listAdapter.toJson(writer, message.getGroupKeyIds());
+    Json.listAdapter.toJson(writer, message.getGroupKeyIds());
     writer.endArray();
   }
 }
