+++ src/main/java/com/streamr/client/rest/StreamrRESTClient.java
@@ -3,9 +3,10 @@ package com.streamr.client.rest;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Types;
 import com.streamr.client.options.StreamrClientOptions;
+import com.streamr.client.protocol.message_layer.Json;
 import com.streamr.client.utils.Address;
-import com.streamr.client.utils.HttpUtils;
 import java.io.IOException;
+import java.net.HttpURLConnection;
 import java.util.List;
 import okhttp3.Call;
 import okhttp3.HttpUrl;
@@ -65,7 +66,7 @@ public abstract class StreamrRESTClient extends AbstractStreamrClient {
     final Call call = client.newCall(request);
     final Response response = call.execute();
     try {
-      HttpUtils.assertSuccessful(response);
+      assertSuccessful(response);
 
       // Deserialize HTTP response to concrete type.
       if (adapter == null) {
@@ -116,7 +117,7 @@ public abstract class StreamrRESTClient extends AbstractStreamrClient {
       HttpUrl url, String requestBody, JsonAdapter<T> adapter, boolean retryIfSessionExpired)
       throws IOException {
     Request.Builder builder =
-        new Request.Builder().url(url).post(RequestBody.create(requestBody, HttpUtils.jsonType));
+        new Request.Builder().url(url).post(RequestBody.create(requestBody, Json.jsonType));
     return executeWithRetry(builder, adapter, retryIfSessionExpired);
   }
 
@@ -156,7 +157,8 @@ public abstract class StreamrRESTClient extends AbstractStreamrClient {
     return post(url, streamJsonAdapter.toJson(stream), streamJsonAdapter);
   }
 
-  public Permission grant(final Stream stream, final Permission.Operation operation, final String user)
+  public Permission grant(
+      final Stream stream, final Permission.Operation operation, final String user)
       throws IOException {
     if (stream == null || operation == null || user == null) {
       throw new IllegalArgumentException("Must give all of stream, operation, and user!");
@@ -168,7 +170,8 @@ public abstract class StreamrRESTClient extends AbstractStreamrClient {
     return post(url, permissionJsonAdapter.toJson(permission), permissionJsonAdapter);
   }
 
-  public Permission grantPublic(final Stream stream, final Permission.Operation operation) throws IOException {
+  public Permission grantPublic(final Stream stream, final Permission.Operation operation)
+      throws IOException {
     if (stream == null || operation == null) {
       throw new IllegalArgumentException("Must give stream and operation!");
     }
@@ -233,5 +236,28 @@ public abstract class StreamrRESTClient extends AbstractStreamrClient {
       builder = builder.addPathSegment(segment);
     }
     return builder.build();
+  }
+
+  /** You might have to close {@code response} if {@code assertSuccessful()} fails. */
+  static void assertSuccessful(Response response) throws IOException {
+    if (!response.isSuccessful()) {
+      String action = response.request().method() + " " + response.request().url().toString();
+
+      switch (response.code()) {
+        case HttpURLConnection.HTTP_NOT_FOUND:
+          throw new ResourceNotFoundException(action);
+        case HttpURLConnection.HTTP_UNAUTHORIZED:
+          throw new AuthenticationException(action);
+        case HttpURLConnection.HTTP_FORBIDDEN:
+          throw new PermissionDeniedException(action);
+        default:
+          throw new RuntimeException(
+              action
+                  + " failed with HTTP status "
+                  + response.code()
+                  + ":"
+                  + response.body().string());
+      }
+    }
   }
 }
