+++ src/test/groovy/com/streamr/client/protocol/message_layer/StreamrSpecification.groovy
@@ -1,7 +1,6 @@
 package com.streamr.client.protocol.message_layer
 
 import com.streamr.client.utils.Address
-import com.streamr.client.utils.HttpUtils
 import spock.lang.Specification
 
 /**
@@ -35,6 +34,6 @@ class StreamrSpecification extends Specification {
     }
 
     protected StreamMessage createMessage(long timestamp = 0, long sequenceNumber = 0, Long previousTimestamp = null, Long previousSequenceNumber = null, Address publisherId = new Address("publisherId"), Map content = [:], String msgChainId = "msgChainId") {
-        return new StreamMessage.Builder().withMessageId(new MessageID("streamId", 0, timestamp, sequenceNumber, publisherId, msgChainId)).withPreviousMessageRef((previousTimestamp != null ? new MessageRef(previousTimestamp, previousSequenceNumber ?: 0) : null)).withSerializedContent(HttpUtils.mapAdapter.toJson(content)).createStreamMessage()
+        return new StreamMessage.Builder().withMessageId(new MessageID("streamId", 0, timestamp, sequenceNumber, publisherId, msgChainId)).withPreviousMessageRef((previousTimestamp != null ? new MessageRef(previousTimestamp, previousSequenceNumber ?: 0) : null)).withSerializedContent(Json.mapAdapter.toJson(content)).createStreamMessage()
     }
 }
